requirements.txt
annotated-types==0.7.0
cachetools==5.5.0
certifi==2024.12.14
charset-normalizer==3.4.1
google-ai-generativelanguage==0.6.10
google-api-core==2.24.0
google-api-python-client==2.157.0
google-auth==2.37.0
google-auth-httplib2==0.2.0
google-generativeai==0.8.3
googleapis-common-protos==1.66.0
grpcio==1.69.0
grpcio-status==1.69.0
httplib2==0.22.0
idna==3.10
proto-plus==1.25.0
protobuf==5.29.2
pyasn1==0.6.1
pyasn1_modules==0.4.1
pydantic==2.10.4
pydantic_core==2.27.2
pyparsing==3.2.1
python-dotenv==1.0.1
requests==2.32.3
rsa==4.9
tqdm==4.67.1
typing_extensions==4.12.2
uritemplate==4.1.1
urllib3==2.3.0
